+++ ecmascript/transforms/compat/tests/es2015_function_name.rs
@@ -228,106 +228,6 @@ export const y = function y() {};
 //
 //"#);
 
-// function_name_export_default_arrow_renaming_module_amd
-test!(
-    ignore,
-    syntax(),
-    |_| chain!(
-        function_name(),
-        shorthand(),
-        arrow(),
-        amd(Default::default())
-    ),
-    function_name_export_default_arrow_renaming_module_amd,
-    r#"
-export default (a) => {
-  return { a() { return a } };
-}
-
-"#,
-    r#"
-define(["exports"], function (_exports) {
-  "use strict";
-
-  Object.defineProperty(_exports, "__esModule", {
-    value: true
-  });
-  _exports.default = void 0;
-
-  var _default = function _default(_a) {
-    return {
-      a: function a() {
-        return _a;
-      }
-    };
-  };
-
-  _exports.default = _default;
-});
-
-"#
-);
-
-// function_name_object
-test!(
-    ignore,
-    syntax(),
-    |_| chain!(
-        resolver(),
-        function_name(),
-        classes(),
-        decorators(decorators::Config {
-            legacy: true,
-            ..Default::default()
-        })
-    ),
-    function_name_object,
-    r#"
-var obj = {
-  f: function () {
-    (function f() {
-      console.log(f);
-    })();
-  },
-
-  h: function () {
-    console.log(h);
-  },
-
-  m: function () {
-    doSmth();
-  }
-};
-
-"#,
-    r#"
-var obj = {
-  f: function f() {
-    (function f() {
-      console.log(f);
-    })();
-  },
-  h: function (_h) {
-    function h() {
-      return _h.apply(this, arguments);
-    }
-
-    h.toString = function () {
-      return _h.toString();
-    };
-
-    return h;
-  }(function () {
-    console.log(h);
-  }),
-  m: function m() {
-    doSmth();
-  }
-};
-
-"#
-);
-
 // function_name_export
 test!(
     // not important
